Get clearer about rejecting in-use objects
As the rules around rejecting non-pending objects are better spelled out now, the
"Rejection" section has now been moved to being a note of the "pending" property.

Also, as FetchObservers are really only associated with *the same fetch as a FetchController*,
not the FetchController itself, the relationship between the two has been relaxed somewhat:
pre-associated FetchObservers and FetchControllers can be swapped around after construction
(though there should be nor real reason to do that).